chore: move policy to prividium
