+++ packages/api/src/log/log.tokens.ts
@@ -1,3 +1,1 @@
 export const LOG_VISIBILITY_POLICY = Symbol("LOG_VISIBILITY_POLICY");
-
-export { NoopLogVisibilityPolicy, RuleBasedLogVisibilityPolicy, LogVisibilityPolicy } from "./log.visibility-policy";
